Fixed tau leaping solver test and added it to integration testing.
