+++ test/test_tau_leaping_solver.py
@@ -14,15 +14,53 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import os
+import sys
+
+sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
+
 import unittest
 import numpy as np
-from example_models import build_example
-from gillespy2 import TauLeapingSolver
-
+from example_models import build_example, build_example_no_tspan
+from gillespy2 import Results, TauLeapingSolver
+from gillespy2.core.gillespyError import SimulationError
 
 class TestBasicTauLeapingSolver(unittest.TestCase):
-    model = build_example()
-    
+
+    def setUp(self):
+        self.model = build_example()
+        self.model_no_tspan = build_example_no_tspan()
+        self.solver = TauLeapingSolver(model=self.model)
+        self.solver_no_tspan = TauLeapingSolver(model=self.model_no_tspan)
+
+    def test_return_type(self):
+        results = self.model.run(
+            solver=self.solver, number_of_trajectories=1
+        ).to_array()
+        labels_results = self.model.run(solver=self.solver, number_of_trajectories=1)
+        self.assertIsInstance(results, np.ndarray)
+        self.assertIsInstance(results[0], np.ndarray)
+        self.assertIsInstance(results[0][0], np.ndarray)
+        self.assertIsInstance(results[0][0][0], float)
+
+    def test_return_type_show_labels(self):
+        results = self.model.run(
+            solver=self.solver, number_of_trajectories=1
+        ).to_array()
+        labels_results = self.model.run(solver=self.solver, number_of_trajectories=1)
+        self.assertIsInstance(labels_results, Results)
+        self.assertIsInstance(labels_results['Sp'], np.ndarray)
+        self.assertIsInstance(labels_results['Sp'][0], float)
+
+    def test_run_example__with_increment_only(self):
+        results = self.solver_no_tspan.run(increment=0.2)
+
+    def test_run_example__with_tspan_only(self):
+        results = self.solver.run()
+
+    def test_run_example__with_tspan_and_increment(self):
+        with self.assertRaises(SimulationError):
+            results = self.solver.run(increment=0.2)
 
 if __name__ == '__main__':
     unittest.main()
